fix: move stage logic during drag
